requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.0.0
reportlab>=4.0.0
